FDT error-proofed Starling. There are a couple of coding quirks in the Starling AS3 that FDT doesn't like, which are removed.
In starling...Texture, classes that return a new starling...Texture use ":Texture", but there's also an import for flash.stage3D...Texture, which confuses FDT since it expects if two Texture classes are imported, both will be referenced with a fully-qualified classname.

FDT also doesn't recognise the correctness of passing in Array.CONSTANTS into a Vector's sort function (because the sort function's signature definition is incorrect). I've wrapped the one case of this in /*FDT_IGNORE*/ tags to prevent this error showing up.